requirements.txt
requests
beautifulsoup4
google-genai
playwright
python-dotenv
